Conflicts:
	app/code/core/Mage/Compiler/Block/Process.php
	app/code/core/Mage/Compiler/Helper/Data.php
	app/code/core/Mage/Compiler/Model/Process.php
	app/code/core/Mage/Compiler/controllers/Adminhtml/Compiler/ProcessController.php
	app/code/core/Mage/Compiler/controllers/ProcessController.php
	app/code/core/Mage/Compiler/etc/adminhtml.xml
	app/code/core/Mage/Compiler/etc/compilation.xml
	app/code/core/Mage/Compiler/etc/config.xml
	app/code/core/Mage/Sales/doc/test.php
	app/design/adminhtml/default/default/layout/compiler.xml
	app/design/adminhtml/default/default/template/compiler/process.phtml
	app/etc/modules/Mage_Compiler.xml
	shell/compiler.php

+++ app/code/core/Mage/Adminhtml/controllers/Promo/CatalogController.php
@@ -20,7 +20,7 @@
  *
  * @category    Mage
  * @package     Mage_Adminhtml
- * @copyright  Copyright (c) 2006-2018 Magento, Inc. (http://www.magento.com)
+ * @copyright  Copyright (c) 2006-2019 Magento, Inc. (http://www.magento.com)
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -39,6 +39,17 @@ class Mage_Adminhtml_Promo_CatalogController extends Mage_Adminhtml_Controller_A
      * @var string
      */
     protected $_dirtyRulesNoticeMessage;
+
+    /**
+     * Controller predispatch method
+     *
+     * @return Mage_Adminhtml_Controller_Action
+     */
+    public function preDispatch()
+    {
+        $this->_setForcedFormKeyActions('delete');
+        return parent::preDispatch();
+    }
 
     protected function _initAction()
     {
@@ -194,6 +205,13 @@ class Mage_Adminhtml_Promo_CatalogController extends Mage_Adminhtml_Controller_A
             try {
                 $model = Mage::getModel('catalogrule/rule');
                 $model->load($id);
+                if (!$model->getRuleId()) {
+                    Mage::getSingleton('adminhtml/session')->addError(
+                        Mage::helper('catalogrule')->__('Unable to find a rule to delete.')
+                    );
+                    $this->_redirect('*/*/');
+                    return;
+                }
                 $model->delete();
                 Mage::getModel('catalogrule/flag')->loadSelf()
                     ->setState(1)
